move TorServicePrefs to package prefs
