+++ topl-service/src/test/java/io/matthewnelson/topl_service/util/TorServicePrefsUnitTest.kt
@@ -1,6 +1,7 @@
 package io.matthewnelson.topl_service.util
 
 import androidx.test.core.app.ApplicationProvider
+import io.matthewnelson.topl_service.prefs.TorServicePrefs
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
